Updated code. added comments and made minor changes to xml

+++ app/src/main/java/com/example/assign2mobile/firstFragment.java
@@ -30,38 +30,30 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link firstFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
+
 public class firstFragment extends Fragment {
 
 
-    private FragmentFirstBinding binding;
-
-    Adapter adapter;
-
-    databaseHelper addrDatabase;
-
-    List<addrObj> allAddr;
-
-    TextView noAddr;
+    private FragmentFirstBinding binding; //defining a view binding to use.
+    Adapter adapter; // defining an adapter to be user to set to recycler view.
+    databaseHelper addrDatabase; //defining a database helper to use.
+    List<addrObj> allAddr; //defining list for all addresses
+    TextView noAddr; //defining xml elements
     RecyclerView recyclerView;
 
-    public firstFragment() {
-        // Required empty public constructor
+    public firstFragment() { // Required empty public constructor
+
     }
 
 
-    public static firstFragment newInstance(String param1, String param2) {
+    public static firstFragment newInstance(String param1, String param2) { //unused, required for fragment.
         firstFragment fragment = new firstFragment();
         Bundle args = new Bundle();
 
         fragment.setArguments(args);
         return fragment;
     }
-    private void displayList(List<addrObj> allAddrs) {
+    private void displayList(List<addrObj> allAddrs) { //function to set the adapter for the recycler view so that the addresses can be displayed
         recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         adapter = new Adapter(this,getActivity(),allAddrs);
         recyclerView.setAdapter(adapter);
@@ -70,13 +62,12 @@ public class firstFragment extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentFirstBinding.inflate(inflater, container, false);
+        binding = FragmentFirstBinding.inflate(inflater, container, false); //defining the view binding.
         return binding.getRoot();
     }
 
@@ -84,21 +75,21 @@ public class firstFragment extends Fragment {
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        addrDatabase = new databaseHelper(getActivity());
+        addrDatabase = new databaseHelper(getActivity()); //creating DB helper to use
 
-        noAddr = binding.noAddrtext;
-        noAddr.setVisibility(View.GONE);
-        allAddr = addrDatabase.getAlladdr();
+        noAddr = binding.noAddrtext; //defining xml elements to variables
         recyclerView = binding.showAddr;
+        noAddr.setVisibility(View.GONE); //hiding one of the elements
+        allAddr = addrDatabase.getAlladdr(); //sending query to DB to get all entries and to return a list
 
-        if(allAddr.isEmpty()){
+        if(allAddr.isEmpty()){ //if the DB is empty, read from the coordinate JSON file and add addresses to DB and the to return the list again.
             addrDatabase.deleteTable();
             addrDatabase.addAddrs(addAddrList());
             allAddr = addrDatabase.getAlladdr();
         }
-        displayList(allAddr);
+        displayList(allAddr); //calling function to setup the recycler view to display the addresses
 
-        binding.addAddrbutton.setOnClickListener(new View.OnClickListener() {
+        binding.addAddrbutton.setOnClickListener(new View.OnClickListener() { //button on click to navigate to the second fragment (add address page)
             @Override
             public void onClick(View view) {
                 NavHostFragment.findNavController(firstFragment.this)
@@ -106,7 +97,7 @@ public class firstFragment extends Fragment {
             }
         });
 
-        binding.searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+        binding.searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() { //creating on query listner to create search functionality on the main page.
             @Override
             public boolean onQueryTextSubmit(String query) {
                 return false;
@@ -115,8 +106,8 @@ public class firstFragment extends Fragment {
             @Override
             public boolean onQueryTextChange(String newText) {
                 // inside on query text change method we are
-                // calling a method to filter our recycler view.
-                if(!allAddr.isEmpty()) {
+                // calling a method to filter the recycler view.
+                if(!allAddr.isEmpty()) { //As long as the address list is not empty, attempt to filter the list based on user criteria
                     filter(newText);
                 }
                 return false;
@@ -125,33 +116,34 @@ public class firstFragment extends Fragment {
 
     }
 
-    public List<addrObj> addAddrList(){
+    public List<addrObj> addAddrList(){ //function to add the address to the DB
         List<Address> addresses;
         double lat,longV;
-        Geocoder geocoder = new Geocoder(requireActivity(), Locale.getDefault());
-        List<addrObj> allAddrs = new ArrayList<>();
+
+        Geocoder geocoder = new Geocoder(requireActivity(), Locale.getDefault()); //creating a geocode to get an address from the lat and long inputs
+        List<addrObj> allAddrs = new ArrayList<>(); //creating a list to store all the addresses
         try {
             JSONObject obj = new JSONObject(loadJSONFromAsset());
-            JSONArray jsonarray = obj.getJSONArray("coordinates");
-            for (int i = 0; i < jsonarray.length(); i++) {
-                JSONObject jsonobject = jsonarray.getJSONObject(i);
+            JSONArray jsonarray = obj.getJSONArray("coordinates"); //loading a JSON array from the input file
+            for (int i = 0; i < jsonarray.length(); i++) { //iterate through the file to get all the coordinate values and addresses
+                 JSONObject jsonobject = jsonarray.getJSONObject(i);
                  lat = Double.parseDouble(jsonobject.getString("latitude"));
-                 longV = Double.parseDouble(jsonobject.getString("longitude"));
+                 longV = Double.parseDouble(jsonobject.getString("longitude")); //getting lat and long values from file
 
-                addresses = geocoder.getFromLocation(lat, longV, 1);
+                 addresses = geocoder.getFromLocation(lat, longV, 1); //search for the address that corresponds to associated coordinates.
 
-                String address = addresses.get(0).getAddressLine(0).trim();
+                 String address = addresses.get(0).getAddressLine(0).trim(); //getting address result from the search
 
-                allAddrs.add(new addrObj(address,lat,longV));
+                 allAddrs.add(new addrObj(address,lat,longV)); //creating a new address object and adding it to the list
             }
-            Toast.makeText(getActivity(), "Added Addresses to list from file", Toast.LENGTH_SHORT).show();
-            return allAddrs;
+            Toast.makeText(getActivity(), "Added Addresses to list from file", Toast.LENGTH_SHORT).show(); //letting user know file was read
+            return allAddrs; //return the list of addresses
         } catch (IOException | JSONException e) {
             throw new RuntimeException(e);
         }
     }
 
-    public String loadJSONFromAsset() throws IOException {
+    public String loadJSONFromAsset() throws IOException { //function to read the json file
         String json;
         InputStream is;
         try {
@@ -168,16 +160,16 @@ public class firstFragment extends Fragment {
         return json;
     }
 
-    private void filter(String text) {
+    private void filter(String text) { //function to search through list of address based on search criteria and to then update the list
         List<addrObj> filteredList = new ArrayList<>();
 
-        for (addrObj item : allAddr) {
+        for (addrObj item : allAddr) { //searching the list for the input
             if (item.getAddress().toLowerCase().contains(text.toLowerCase())) {
                 filteredList.add(item);
             }
         }
 
-        adapter.filterList(filteredList);
+        adapter.filterList(filteredList); //filtering list based on criteria
 
     }
 }
